- [x] Add ML zip models to Hubbup.Web - [x] Conclude changes in CreateMikLabelModel - [x] moves shared code to Hubbup.MikLabelModel
